✨ use case: get ranking
