+++ packages/core/src/ranking/use-cases/GetLastWeekRankingWinnersUseCase.ts
@@ -1,9 +1,13 @@
-import { User } from '../../global/domain/entities'
-import type { RankingUserDto, TierDto } from '../domain/entities/dtos'
-import { Tier } from '../domain/entities'
-import type { UserDto } from '#profile/domain/entities/dtos/UserDto'
-import type { RankingService } from '../interfaces'
+import { OrdinalNumber } from '#global/domain/structures/OrdinalNumber'
 import type { UseCase } from '#global/interfaces/UseCase'
+import type { RankingUserDto, TierDto } from '../domain/entities/dtos'
+import type { RankersRepository, TiersRepository } from '../interfaces'
+import { TierNotFoundError } from '../domain/errors'
+
+type Request = {
+  lastWeekTierPosition: number
+  currentWeekTierPosition: number
+}
 
 type Response = Promise<{
   isUserLoser: boolean
@@ -11,44 +15,36 @@ type Response = Promise<{
   lastWeekRankingWinners: RankingUserDto[]
 }>
 
-export class GetLastWeekRankingWinnersUseCase implements UseCase<UserDto, Response> {
-  constructor(private rankingsService: RankingService) {}
+export class GetLastWeekRankingWinnersUseCase implements UseCase<Request, Response> {
+  constructor(
+    private readonly tiersRepository: TiersRepository,
+    private readonly rankersRepository: RankersRepository,
+  ) {}
 
-  async execute(userDto: UserDto) {
-    const user = User.create(userDto)
-    const lastWeekTierPosition = await this.fetchLastWeekTierPosition(user)
-    const lastWeekTier = await this.fetchLastWeekTier(lastWeekTierPosition)
-    const lastWeekRankingWinners = await this.fetchLastWeekRankingWinners(
-      lastWeekTier.id.value,
+  async execute({
+    lastWeekTierPosition: lastWeekTierPositionValue,
+    currentWeekTierPosition: currentWeekTierPositionValue,
+  }: Request) {
+    const lastWeekTierPosition = OrdinalNumber.create(lastWeekTierPositionValue)
+    const currentWeekTierPosition = OrdinalNumber.create(currentWeekTierPositionValue)
+    const lastWeekTier = await this.findLastWeekTier(lastWeekTierPosition)
+    const lastWeekRankingWinners = await this.rankersRepository.findAllByTier(
+      lastWeekTier.id,
     )
-    const isUserLoser = user.tier.position.number.isLessThan(lastWeekTier.position.number)
+    const isUserLoser = currentWeekTierPosition.number.isLessThan(
+      lastWeekTier.position.number,
+    )
 
     return {
       isUserLoser: isUserLoser.isTrue,
       lastWeekTier: lastWeekTier.dto,
-      lastWeekRankingWinners,
+      lastWeekRankingWinners: lastWeekRankingWinners.map((ranker) => ranker.dto),
     }
   }
 
-  private async fetchLastWeekTier(tierPosition: number) {
-    const response = await this.rankingsService.fetchTierByPosition(tierPosition)
-    if (response.isFailure) response.throwError()
-    return Tier.create(response.body)
-  }
-
-  private async fetchLastWeekRankingWinners(lastWeekTierId: string) {
-    const response = await this.rankingsService.fetchRankingWinnersByTier(lastWeekTierId)
-    if (response.isSuccessful) {
-      return response.body
-    }
-    return []
-  }
-
-  private async fetchLastWeekTierPosition(user: User) {
-    const response = await this.rankingsService.fetchLastWeekRankingUserTierPosition(
-      user.id.value,
-    )
-    if (response.isFailure) return user.tier.position.value
-    return response.body.position
+  private async findLastWeekTier(tierPosition: OrdinalNumber) {
+    const tier = await this.tiersRepository.findByPosition(tierPosition)
+    if (!tier) throw new TierNotFoundError()
+    return tier
   }
 }
