+++ packages/core/src/ranking/use-cases/UpdateRankingsUseCase.ts
@@ -1,106 +1,80 @@
-import { Tier, type RankingUser } from '../domain/entities'
+import type { Id } from '#global/domain/structures/Id'
+import type { EventBroker } from '#global/interfaces/EventBroker'
+import type { UseCase } from '#global/interfaces/UseCase'
+import type { Tier } from '../domain/entities'
+import {
+  RankingLosersDefinedEvent,
+  RankingUpdatedEvent,
+  RankingWinnersDefinedEvent,
+} from '../domain/events'
 import { Ranking } from '../domain/structures'
-import type { RankingService } from '../interfaces'
+import type { RankersRepository, TiersRepository } from '../interfaces'
 
-export class UpdateRankingsUseCase {
-  constructor(private readonly rankingService: RankingService) {}
+export class UpdateRankingsUseCase implements UseCase<void, void> {
+  constructor(
+    private readonly tiersRepository: TiersRepository,
+    private readonly rankersRepository: RankersRepository,
+    private readonly eventBroker: EventBroker,
+  ) {}
 
   async execute() {
     const [tiers] = await Promise.all([
-      this.fetchTiers(),
-      this.deleteLastWeekRankingUsers(),
-      this.updateLastWeekRankingPositions(),
+      this.tiersRepository.findAll(),
+      this.rankersRepository.removeAll(),
     ])
-    await this.saveLosers(tiers)
-    await this.saveWinners(tiers)
-    await this.rankingService.allowUsersSeeRankingResult()
-    await this.resetRankingUsersXp()
+    await this.addLosers(tiers)
+    await this.addWinners(tiers)
+
+    for (const tier of tiers)
+      await this.eventBroker.publish(new RankingUpdatedEvent({ tierId: tier.id.value }))
   }
 
-  private async saveLosers(tiers: Tier[]) {
+  private async addLosers(tiers: Tier[]) {
     for (let index = 0; index < tiers.length; index++) {
       const currentTier = tiers[index]
       if (!currentTier) continue
       let previousTier = tiers[index - 1]
       previousTier = !previousTier ? currentTier : previousTier
 
-      const ranking = await this.fetchRankingByTier(currentTier.id.value)
+      const ranking = await this.findRankingByTier(currentTier.id)
       const losers = ranking.losers
 
-      const response = await this.rankingService.saveRankingLosers(
-        losers,
-        currentTier.id.value,
-      )
+      await this.rankersRepository.addLosers(losers, currentTier.id)
 
-      if (response.isFailure) {
-        await this.deleteLastWeekRankingUsers()
-        response.throwError()
-      }
-
-      if (!previousTier.isEqualTo(currentTier)) {
-        await this.updateRankingUsersTier(losers, previousTier.id.value)
+      if (previousTier.isEqualTo(currentTier).isFalse) {
+        const event = new RankingLosersDefinedEvent({
+          tierId: previousTier.id.value,
+          losersIds: losers.map((loser) => loser.id.value),
+        })
+        await this.eventBroker.publish(event)
       }
     }
   }
 
-  private async saveWinners(tiers: Tier[]) {
+  private async addWinners(tiers: Tier[]) {
     for (let index = tiers.length - 1; index >= 0; index--) {
       const currentTier = tiers[index]
       if (!currentTier) continue
       let nextTier = tiers[index + 1]
       nextTier = !nextTier ? currentTier : nextTier
 
-      const ranking = await this.fetchRankingByTier(currentTier.id.value)
+      const ranking = await this.findRankingByTier(currentTier.id)
       const winners = ranking.winners
 
-      const response = await this.rankingService.saveRankingWinners(
-        winners,
-        currentTier.id.value,
-      )
+      await this.rankersRepository.addWinners(winners, currentTier.id)
 
-      if (response.isFailure) {
-        await this.deleteLastWeekRankingUsers()
-        response.throwError()
-      }
-
-      if (!nextTier.isEqualTo(currentTier)) {
-        await this.updateRankingUsersTier(winners, nextTier.id.value)
+      if (nextTier.isEqualTo(currentTier).isFalse) {
+        const event = new RankingWinnersDefinedEvent({
+          tierId: nextTier.id.value,
+          winnersIds: winners.map((winner) => winner.id.value),
+        })
+        await this.eventBroker.publish(event)
       }
     }
   }
 
-  private async fetchTiers() {
-    const tiersResponse = await this.rankingService.fetchTiers()
-    if (tiersResponse.isFailure) tiersResponse.throwError()
-    return tiersResponse.body.map(Tier.create)
-  }
-
-  private async fetchRankingByTier(tierId: string) {
-    const response = await this.rankingService.fetchLastWeekRankingUsersByTier(tierId)
-    if (response.isFailure) response.throwError()
-    return Ranking.create(response.body)
-  }
-
-  private async deleteLastWeekRankingUsers() {
-    const response = await this.rankingService.deleteLastWeekRankingUsers()
-    if (response.isFailure) response.throwError()
-  }
-
-  private async resetRankingUsersXp() {
-    const response = await this.rankingService.resetRankingUsersXp()
-    if (response.isFailure) response.throwError()
-  }
-
-  private async updateLastWeekRankingPositions() {
-    const response = await this.rankingService.updateLastWeekRankingPositions()
-    if (response.isFailure) response.throwError()
-  }
-
-  private async updateRankingUsersTier(rankingUsers: RankingUser[], tierId: string) {
-    const response = await this.rankingService.updateRankingUsersTier(
-      rankingUsers,
-      tierId,
-    )
-    if (response.isFailure) response.throwError()
+  private async findRankingByTier(tierId: Id) {
+    const rankers = await this.rankersRepository.findAllByTier(tierId)
+    return Ranking.create(rankers.map((ranker) => ranker.dto))
   }
 }
